requirements.txt
alembic==1.4.3
altgraph==0.17
aniso8601==8.0.0
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
auto-py-to-exe==2.7.8
autopep8==1.5.4
Babel==2.8.0
bcrypt==3.2.0
bottle==0.12.18
bottle-websocket==0.2.9
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
distlib==0.3.1
Django==3.1.2
ecdsa==0.14.1
Eel==0.12.4
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.3
future==0.18.2
gevent==20.9.0
gevent-websocket==0.10.1
greenlet==0.4.17
gunicorn==20.0.4
idna==2.10
isort==5.5.4
itsdangerous==1.1.0
Jinja2==2.11.2
jose==1.0.0
jwt==1.0.0
lazy-object-proxy==1.4.3
loguru==0.5.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
paramiko==2.7.2
pefile==2019.4.18
pep8==1.7.1
postgres==3.0.0
prettytable==0.7.2
psycopg2==2.8.6
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.8
pyflakes==2.2.0
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.8
pylint==2.6.0
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.2.0
pytz==2020.1
pywin32-ctypes==0.2.0
requests==2.24.0
requests-toolbelt==0.9.1
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
sage-testsuite==0.1.0a6
six==1.15.0
SQLAlchemy==1.3.4
sqlparse==0.4.1
sshtunnel==0.1.5
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
urlopen==1.0.0
virtualenv==20.0.34
Werkzeug==1.0.1
whichcraft==0.6.1
win32-setctime==1.0.3
wrapt==1.12.1
WTForms==2.3.3
zope.event==4.5.0
zope.interface==5.1.2
